Cleaning to clean better.

+++ bench/BMRotation.cpp
@@ -13,9 +13,9 @@ struct BMRotation : public benchmark::Fixture
     {
       io::imread(filename, m_input);
 
-      int nr = m_input.domain().shape()[0];
-      int nc = m_input.domain().shape()[1];
-      m_output = image2d<rgb8>(box2d{ {-nc+1,0}, {1,nr}});
+      short nr = m_input.domain().shape()[0];
+      short nc = m_input.domain().shape()[1];
+      m_output = image2d<rgb8>(box2d{ {(short)(-nc+1),0}, {1,nr}});
       m_bytes = nr * nc * sizeof(rgb8);
     }
 
@@ -37,7 +37,7 @@ void rotate_ptr(const image2d<T>& f, image2d<T>& out)
   const T* ptr_in = & f({0,0});
 
   for (int i = 0; i < nr; ++i) {
-    T* ptr_out = &out({0,i});
+    T* ptr_out = &out({0,(short)i});
     for (int j = 0; j < nc; ++j) {
       *ptr_out = ptr_in[j];
       ptr_out = (T*)((char*)ptr_out - ostride);
@@ -51,7 +51,7 @@ __attribute__ ((noinline))
 void rotate_p(const image2d<T>& f, image2d<T>& out)
 {
   mln_foreach(point2d p, f.domain()) {
-    point2d q{-p[1], p[0]};
+    point2d q{(short)(-p[1]), p[0]};
     out(q) = f(p);
   }
 }
@@ -62,7 +62,7 @@ __attribute__ ((noinline))
 void rotate_pix(const image2d<T>& f, image2d<T>& out)
 {
   mln_foreach(auto px, f.pixels()) {
-    point2d q{-px.point()[1], px.point()[0]};
+    point2d q{(short)(-px.point()[1]), px.point()[0]};
     out(q) = px.val();
   }
 }
